🐛 Allowed possible `posts/` prefix to detect URL for draft posts
no issue

When processing a Medium export as provided from the Settings menu, the posts to convert are in a subfolder called `posts`. The current post processing would run into an error when trying to process a `post.url` that simply doesn't exist for a draft post.

This is because when the post is stored in a subfolder, the regex that is used to differentiate between a draft and a published post, and therefore detect the correct post URL, didn't check for a `post/` prefix.

This commit adds the optional prefix, so the name can either start with `posts/draft_whateverhere`, or with `draft_whateverhere`.