✨ Handle linked images for WordPress
no issue

Koenig has no concept of linked images, other than using an HTML card. This change adds `kg-card` classes to the anchor tag that contains an image and wraps it in an HTML card. We can't convert the image fully to a kg image (wrapped in a `figure` tag), as the Mobiledoc parser doesn't allow it within the anchor tag, even when wrapped in an HTML card.

This should probably be fixed within the Mobiledoc parser, but is ok for now.