Add option to use og:image as feature image
Substack does not let you set a feature image as such, instead letting you pick a social image, which is also used on archive pages.

This is a more sensible way to handle feature images, allowing a migration to mirror Substack better.

The code and test to use the first image has been removed.