md-jekyll-export: Add support Jekyll posts in HTML format.
HTML posts are detected using an ".html" file extension.

Before, the entire contents of Jekyll posts was referred to as
"Markdown". This was not completely accurate, as it is a combination of "front
matter" plus Markdown. And this not accurate at all ilf the file content
was actually html. Now, we use "fileContents" to refer to the whole
file.

Before, the front matter was being parsed twice, once during metadata
processing and a second time for "markdown" processing. Now, frontmatter
only gets processed ones, which should be a small speed-up.

Before, "process-content.js" was almost exclusively updating HTML and
not any other type of content. Now that HTML support has been added,
that's the only thing that it does, so it renamed "process-html.js" to
be more clear.

I'm not sure if Jekyll supports multiple extensions for HTML files like
it does for Markdown files. Here only support for ".html" is added. If
an unknown file type is detected, an error will be thrown.

Fixes #507