mg-jekyll-export: fix multiple problems with reading the zip.
First, the intent was to read files in the `_drafts/` folder was well,
but the Regex was not matching it.

Second, the intent was to match against the path stored in the the
zipfile, but that was broken. Confusingly, "entryName" holds just the
file name while "zipEntry.entryName" holds the full path. The fix is
to switch "entryName" we use.

Third, only the bare file name was being reported when skipped. It's
more useful to see the full path within the zip, in case the problem is
with the folder structure in there.

Fixes #510