Fixed missing version breaking yarn workspaces
no issue

In order to prevent publishing the mg-stripe-csv, I tried to remove the version from package.json (since setting it to private didn't work). However, this broke yarn workspaces (dependencies were not installed for this package).

To get around this, `yarn ship` now adds the `--no-private` flag to the larna command.